+++ CHANGELOG.md
@@ -29,6 +29,9 @@
 
 ### Changed
 - `bit-gen-reads` fragment size and long-read read lengths are now pulled from a normal distribution rather than uniform
+- `bit-gen-kraken2-tax-plots` changed to `bit-kraken2-tax-plots`
+- `bit-kraken2-to-taxon-summaries` changed to `bit-kraken2-tax-summary`
+- `bit-combine-kraken2-taxon-summaries` changed to `bit-kraken2-combine-tax-summaries`, and modularized
 
 ---